requirements.txt
cached-property
jinja2
pyyaml
setuptools
six
wheel
